test coverage implemented and test files refactored / #3
