+++ tutorials/conditional/conditional_test.go
@@ -1,20 +1,19 @@
-package tests
+package conditional
 
 import (
 	"testing"
-
-	"github.com/abhinna1/LearnGo/tutorials/conditional"
+	// "github.com/abhinna1/LearnGo/tutorials/conditional"
 )
 
 func TestIsEven(t *testing.T) {
 	const oddNumber int = 1
 	const evenNumber int = 2
 
-	if conditional.IsEven(oddNumber) {
+	if IsEven(oddNumber) {
 		t.Errorf("Number is odd")
 	}
 
-	if !conditional.IsEven(evenNumber) {
+	if !IsEven(evenNumber) {
 		t.Errorf("Number is even.")
 	}
 }
@@ -25,10 +24,10 @@ func TestAccessControl(t *testing.T) {
 	const userRole string = "user"
 	const invalidRole string = "invalid"
 
-	ownerAccess := conditional.AccessControl(ownerRole)
-	adminAccess := conditional.AccessControl(adminRole)
-	userAccess := conditional.AccessControl(userRole)
-	invalidAccess := conditional.AccessControl(invalidRole)
+	ownerAccess := AccessControl(ownerRole)
+	adminAccess := AccessControl(adminRole)
+	userAccess := AccessControl(userRole)
+	invalidAccess := AccessControl(invalidRole)
 
 	if ownerAccess != 1 {
 		t.Errorf("Invalid Owner Access.")
@@ -53,10 +52,10 @@ func TestAccessControlSwitchCases(t *testing.T) {
 	const userRole string = "user"
 	const invalidRole string = "invalid"
 
-	ownerAccess := conditional.AccessControlSwitchCases(ownerRole)
-	adminAccess := conditional.AccessControlSwitchCases(adminRole)
-	userAccess := conditional.AccessControlSwitchCases(userRole)
-	invalidAccess := conditional.AccessControlSwitchCases(invalidRole)
+	ownerAccess := AccessControlSwitchCases(ownerRole)
+	adminAccess := AccessControlSwitchCases(adminRole)
+	userAccess := AccessControlSwitchCases(userRole)
+	invalidAccess := AccessControlSwitchCases(invalidRole)
 
 	if ownerAccess != 1 {
 		t.Errorf("Invalid Owner Access.")
