+++ tutorials/loops/loops_test.go
@@ -1,9 +1,7 @@
-package tests
+package loops
 
 import (
 	"testing"
-
-	"github.com/abhinna1/LearnGo/tutorials/loops"
 )
 
 // Function to check if two integer lists are equal.
@@ -21,7 +19,7 @@ func assertLists(list1 []int, list2 []int) bool {
 
 func TestForLoop(t *testing.T) {
 	list := []int{0, 1, 2, 3, 4, 5, 6, 7, 8, 9}
-	var count []int = loops.PrintCounter(10)
+	var count []int = PrintCounter(10)
 	if len(list) != len(count) {
 		t.Errorf("Lists are not equal.")
 	}
