+++ tutorials/variables/variables_test.go
@@ -1,17 +1,15 @@
-package tests
+package variables
 
 import (
 	"testing"
-
-	"github.com/abhinna1/LearnGo/tutorials/variables"
 )
 
 func TestVariables(t *testing.T) {
 	// test const variable.
-	if variables.CONSTANT1 != "" {
+	if CONSTANT1 != "" {
 		t.Errorf("Constant did not match")
 	}
-	if variables.Number_variable != 1 {
+	if Number_variable != 1 {
 		t.Errorf("Variable did not match.")
 	}
 }
